+++ Sources/UtiliKit/Container/ContainerViewController.swift
@@ -10,21 +10,22 @@ import UIKit
 open class ContainerViewController: UIViewController {
     
     //MARK: Properties
-    public var managedChildren: [ManagedChild] = []
-    public private(set) var isTransitioning: Bool = false
-    public var shouldAutomaticallyTransitionOnLoad: Bool = true
-    public var visibleController: UIViewController? {
-        didSet {
-            guard let visible = visibleController else { return }
-            transition(to: visible)
-        }
+    open var managedChildren: [ManagedChild] = []
+    open var visibleIndex: Int? { return visibleController.flatMap(index(ofChild:)) }
+    open var visibleChild: ManagedChild? { return managedChildren.first { $0.viewController === visibleController } }
+    
+    open private(set) var isTransitioning: Bool = false
+    open var shouldAutomaticallyTransitionOnLoad: Bool = true
+    
+    open var visibleController: UIViewController? {
+        didSet { visibleController.map { transition(to: $0) } }
     }
     
-    public var visibleIndex: Int? {
-        return visibleController.flatMap { index(ofChild: $0 ) }
-    }
+    open weak var delegate: ContainerViewControllerDelegate?
     
-    public weak var delegate: ContainerViewControllerDelegate?
+    // MARK: Transitioning
+    private var containerTransitionContext: UIViewControllerContextTransitioning?
+    open var containerTransitionCoordinator: ContainerViewControllerTransitionCoordinator?
     
     //MARK: Initializers
     public convenience init(managedChildren: [ManagedChild], delegate: ContainerViewControllerDelegate? = nil) {
@@ -53,6 +54,10 @@ open class ContainerViewController: UIViewController {
 //MARK: Public Interface
 extension ContainerViewController {
     
+    open func transitionToControllerForChild(withIdentifier identifier: AnyHashable, completion: ((Bool) -> Void)? = nil) {
+        managedChildren.first { identifier == $0.identifier }.map { transitionToController(for: $0, completion: completion) }
+    }
+    
     open func transitionToController(for child: ManagedChild, completion: ((Bool) -> Void)? = nil) {
         if !managedChildren.contains { $0.viewController === child.viewController } {
             managedChildren.insert(child, at: managedChildren.startIndex)
@@ -76,7 +81,6 @@ extension ContainerViewController {
 private extension ContainerViewController {
     
     func transition(to destination: UIViewController, completion: ((Bool) -> Void)? = nil) {
-        
         //Ensure that the view is loaded, we're not already transitioning and the transition will result in a move
         guard isViewLoaded && !isTransitioning, visibleController != destination else { completion?(true); return }
         guard let source = visibleController else {
@@ -85,67 +89,77 @@ private extension ContainerViewController {
             prepareForTransitioning(from: nil, to: destination, animated: false)
             view.addSubview(destination.view)
             configure(destinationView: destination.view, inContainer: view)
-            finishTransitioning(from: nil, to: destination, animated: false)
-            completion?(true)
-            return
+            finishTransitioning(from: nil, to: destination, success: true, animated: false)
+            completion?(true); return
         }
         
         guard delegate?.containerViewController(self, shouldTransitionFrom: source, to: destination) ?? true else { completion?(false); return }
         
-        //Inform the delegate transitioning is set to begin
-        delegate?.containerViewController(self, didBeginTransitioningFrom: source, to: destination)
-        
-        //Begin the internal transitioning process, using a UIViewControllerAnimatedTransitioning object
         prepareForTransitioning(from: source, to: destination, animated: true)
-        let animator = delegate?.containerViewController(self, animationControllerForTransitionFrom: source, to: destination) ?? ContainerViewControllerTransitionAnimator()
-        let transitionContext = ContainerTransitionContext(containerViewController: self, sourceViewController: source, destinationViewController: destination) { [weak self] finished in
-            guard let blockSelf = self else { return }
-            blockSelf.configure(destinationView: destination.view, inContainer: blockSelf.view)
-            blockSelf.finishTransitioning(from: source, to: destination, animated: true)
-            blockSelf.delegate?.containerViewController(blockSelf, didFinishTransitioningFrom: source, to: destination)
-            completion?(finished)
-        }
+        defer { delegate?.containerViewController(self, didBeginTransitioningFrom: source, to: destination) }
         
-        //Instruct the animator to begin transitioning
-        animator.animateTransition(using: transitionContext)
+        let context = configuredTransitionContext(from: source, to: destination)
+        let animator = delegate?.containerViewController(self, animationControllerForTransitionFrom: source, to: destination) ?? DefaultContainerTransitionAnimator()
+        let interactor = delegate?.containerViewController(self, interactionControllerForTransitionFrom: source, to: destination)
+        
+        containerTransitionCoordinator = ContainerTransitionCoordinator(context: context, animator: animator, interactor: interactor)
+        if let interactor = interactor, interactor.wantsInteractiveStart {
+            interactor.startInteractiveTransition(context, using: animator)
+        } else {
+            animator.animateTransition(using: context)
+        }
     }
 }
 
 //MARK: Helper
 private extension ContainerViewController {
     
+    func configuredTransitionContext(from source: UIViewController, to destination: UIViewController) -> ContainerTransitionContext {
+        let context = ContainerTransitionContext(containerView: view, fromViewController: source, toViewController: destination)
+        context.completion = { [weak self] finished in
+            guard let self = self else { return }
+            self.finishTransitioning(from: source, to: destination, success: finished, animated: true)
+            self.delegate?.containerViewController(self, didFinishTransitioningFrom: source, to: destination)
+        }
+        
+        return context
+    }
+    
     func prepareForTransitioning(from source: UIViewController?, to destination: UIViewController, animated: Bool) {
         isTransitioning = true
         
-        #if swift(>=4.2)
-        source?.willMove(toParent: nil)
-        destination.willMove(toParent: self)
+        source?.beginAppearanceTransition(false, animated: animated)
+        destination.beginAppearanceTransition(true, animated: animated)
+        
         addChild(destination)
-        #else
-        source?.willMove(toParentViewController: nil)
-        destination.willMove(toParentViewController: self)
-        addChildViewController(destination)
-        #endif
+        destination.didMove(toParent: self)
     }
     
     func configure(destinationView: UIView, inContainer container: UIView) {
+        //It is the animator's responsibility (as with all UIViewControllerAnimatedTransitioning objects) to add the destinationView as a subview of the container view. The container will simply ensure the proper layout is used once the transition is completed.
+        
         destinationView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         destinationView.frame = container.bounds
     }
     
-    func finishTransitioning(from source: UIViewController?, to destination: UIViewController, animated: Bool) {
-        source?.view.removeFromSuperview()
-        #if swift(>=4.2)
-        source?.removeFromParent()
-        source?.didMove(toParent: nil)
-        destination.didMove(toParent: self)
-        #else
-        source?.removeFromParentViewController()
-        source?.didMove(toParentViewController: nil)
-        destination.didMove(toParentViewController: self)
-        #endif
+    func finishTransitioning(from source: UIViewController?, to destination: UIViewController, success: Bool, animated: Bool) {
+        destination.endAppearanceTransition()
+        source?.endAppearanceTransition()
         
-        visibleController = destination
+        if success {
+            source?.willMove(toParent: nil)
+            source?.removeFromParent()
+           
+            visibleController = destination
+            
+        } else {
+            destination.willMove(toParent: nil)
+            destination.removeFromParent()
+            
+            visibleController = source
+        }
+        
+        containerTransitionContext = nil
         isTransitioning = false
     }
 }